GEN-AI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
python-dotenv==1.0.0
lxml==4.9.3
requests==2.31.0 